desktop_agent/requirements.txt
# Desktop Agent Dependencies
PySide6>=6.5.0
httpx>=0.25.0
mss>=9.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
